requirements.txt
Django~=5.0.7
typing_extensions~=4.12.2
pillow~=10.4.0
pip~=24.1.2
attrs~=23.2.0
filelock~=3.15.4
asgiref~=3.8.1
sqlparse~=0.5.0
psycopg~=3.2.1
psycopg2~=2.9.9
PyYAML~=6.0.2
requests~=2.32.3
djangorestframework~=3.15.2
uritemplate~=4.1.1
inflection~=0.5.1
pytz~=2024.1